+++ app/Http/Controllers/TagController.php
@@ -18,10 +18,26 @@ class TagController extends Controller
      */
     public function index()
     {
-        // get all tags
-        $tags = Tag::all();
-        // return view
-        return view('tags', compact('tags'));
+        $log = new LogController();
+        if (auth()->user()) {
+            if (auth()->user()->hasPermissionTo('view tags')) {
+                // get all tags
+                $tags = Tag::all();
+                $log->logMe("info", "User ID: " . auth()->user()->id . " - " . auth()->user()->name . " viewed tags", "GET", request()->ip());
+                // return view
+                return view('tags', compact('tags'));
+            } else {
+                $log->logMe("warning", "User ID: " . auth()->user()->id . " - " . auth()->user()->name . " tried to view tags", "GET", request()->ip());
+                session()->flash('message', 'You are not allowed to view tags');
+                session()->flash('icon', 'error');
+                return redirect()->back();
+            }
+        } else {
+            $log->logMe("warning", "Guest tried to view tags", "GET", request()->ip());
+            session()->flash('message', 'You must be logged in to view tags');
+            session()->flash('icon', 'error');
+            return redirect()->back();
+        }
     }
 
     /**
@@ -46,19 +62,34 @@ class TagController extends Controller
     public function store(StoreTagRequest $request)
     {
         $log = new LogController();
+        if (auth()->user()) {
+            if (auth()->user()->hasPermissionTo('create tags')) {
 
-        $tag = $request->validated();
 
-        $tag['slug'] = Str::slug($tag['name']);
+                $tag = $request->validated();
 
-        $ta = Tag::create($tag);
+                $tag['slug'] = Str::slug($tag['name']);
 
-        $log->logMe("info", "Created tag ID: $ta->id", "POST", $request->ip());
+                $ta = Tag::create($tag);
 
-        session()->flash('message', 'Tag created successfully');
-        session()->flash('icon', 'success');
+                $log->logMe("info", "Created tag ID: $ta->id", "POST", $request->ip());
 
-        return redirect()->back();
+                session()->flash('message', 'Tag created successfully');
+                session()->flash('icon', 'success');
+
+                return redirect()->back();
+            } else {
+                $log->logMe("warning", "User ID: " . auth()->user()->id . " - " . auth()->user()->name . " tried to create tag", "POST", request()->ip());
+                session()->flash('message', 'You are not allowed to create tags');
+                session()->flash('icon', 'error');
+                return redirect()->back();
+            }
+        } else {
+            $log->logMe("warning", "Guest tried to create tag", "POST", request()->ip());
+            session()->flash('message', 'You must be logged in to create tags');
+            session()->flash('icon', 'error');
+            return redirect()->back();
+        }
     }
 
     /**
@@ -75,11 +106,27 @@ class TagController extends Controller
      */
     public function edit($id)
     {
-        $tag = Tag::find($id);
-        // return ajax response with tag data
-        return response()->json(
-            $tag
-        );
+        $log = new LogController();
+        if (auth()->user()) {
+            if (auth()->user()->hasPermissionTo('view tags')) {
+                $log->logMe("info", "User ID: " . auth()->user()->id . " - " . auth()->user()->name . " viewed tag ID: $id", "GET", request()->ip());
+                $tag = Tag::find($id);
+                // return ajax response with tag data
+                return response()->json(
+                    $tag
+                );
+            } else {
+                $log->logMe("warning", "User ID: " . auth()->user()->id . " - " . auth()->user()->name . " tried to view tag ID: $id", "GET", request()->ip());
+                session()->flash('message', 'You are not allowed to view tags');
+                session()->flash('icon', 'error');
+                return redirect()->back();
+            }
+        } else {
+            $log->logMe("warning", "Guest tried to view tag ID: $id", "GET", request()->ip());
+            session()->flash('message', 'You must be logged in to view tags');
+            session()->flash('icon', 'error');
+            return redirect()->back();
+        }
     }
 
     public function allTags()
@@ -95,15 +142,31 @@ class TagController extends Controller
      */
     public function update(UpdateTagRequest $request, Tag $tag)
     {
-        // update tag
-        $tag->update($request->validated());
-        $ta = Tag::find($tag->id);
         $log = new LogController();
-        $log->logMe("info", "Updated tag ID: $ta->id", "PUT", $request->ip());
-        // redirect to tag index
-        session()->flash('message', 'Tag updated successfully');
-        session()->flash('icon', 'success');
-        return redirect()->back();
+        if (auth()->user()) {
+            if (auth()->user()->hasPermissionTo('edit tags')) {
+                $log->logMe("info", "User ID: " . auth()->user()->id . " - " . auth()->user()->name . " updated tag ID: $tag->id", "PUT", $request->ip());
+                // update tag
+                $tag->update($request->validated());
+                $ta = Tag::find($tag->id);
+                $log = new LogController();
+                $log->logMe("info", "Updated tag ID: $ta->id", "PUT", $request->ip());
+                // redirect to tag index
+                session()->flash('message', 'Tag updated successfully');
+                session()->flash('icon', 'success');
+                return redirect()->back();
+            } else {
+                $log->logMe("warning", "User ID: " . auth()->user()->id . " - " . auth()->user()->name . " tried to update tag ID: $tag->id", "PUT", $request->ip());
+                session()->flash('message', 'You are not allowed to update tags');
+                session()->flash('icon', 'error');
+                return redirect()->back();
+            }
+        } else {
+            $log->logMe("warning", "Guest tried to update tag ID: $tag->id", "PUT", $request->ip());
+            session()->flash('message', 'You must be logged in to update tags');
+            session()->flash('icon', 'error');
+            return redirect()->back();
+        }
     }
 
     /**
@@ -111,15 +174,28 @@ class TagController extends Controller
      */
     public function destroy(Tag $tag, Request $request)
     {
-        // delete tag
-        $tag->delete();
-
         $log = new LogController();
+        if (auth()->user()) {
+            if (auth()->user()->hasPermissionTo('delete tags')) {
+                // delete tag
+                $tag->delete();
 
-        $log->logMe("warning", "Deleted tag ID: $tag->id", "DELETE", $request->ip());
-        session()->flash('message', 'Tag deleted successfully');
-        session()->flash('icon', 'success');
-        // redirect to tag index
-        return redirect()->back();
+                $log->logMe("warning", "Deleted tag ID: $tag->id", "DELETE", $request->ip());
+                session()->flash('message', 'Tag deleted successfully');
+                session()->flash('icon', 'success');
+                // redirect to tag index
+                return redirect()->back();
+            } else {
+                $log->logMe("warning", "User ID: " . auth()->user()->id . " - " . auth()->user()->name . " tried to delete tag ID: $tag->id", "DELETE", $request->ip());
+                session()->flash('message', 'You are not allowed to delete tags');
+                session()->flash('icon', 'error');
+                return redirect()->back();
+            }
+        } else {
+            $log->logMe("warning", "Guest tried to delete tag ID: $tag->id", "DELETE", $request->ip());
+            session()->flash('message', 'You must be logged in to delete tags');
+            session()->flash('icon', 'error');
+            return redirect()->back();
+        }
     }
 }
